AV-180281 : Add support for Cilium CNI in Kubernetes with AKO
