+++ internal/lib/constants.go
@@ -24,6 +24,7 @@ const (
 	NCP_CNI                   = "ncp"
 	OPENSHIFT_CNI             = "openshift"
 	OVN_KUBERNETES_CNI        = "ovn-kubernetes"
+	CILIUM_CNI                = "cilium"
 	INGRESS_API               = "INGRESS_API"
 	AviConfigMap              = "avi-k8s-config"
 	AviSecret                 = "avi-secret"
